# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant Billing API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
}

export interface MenuItem {
  id: number;
  name: string;
  price: number;
  category: string;
  imageUrl: string;
  available: boolean;
  createdAt: string;
}

export interface CreateMenuItemBody {
  name: string;
  price: number;
  category: string;
  imageUrl: string;
  available?: boolean;
}

export interface OrderItem {
  menuItemId: number;
  menuItemName: string;
  quantity: number;
  unitPrice: number;
  subtotal: number;
}

export interface Order {
  id: number;
  orderNumber: string;
  totalAmount: number;
  paymentMethod: string;
  status: string;
  items: OrderItem[];
  createdAt: string;
}

export type CreateOrderBodyItemsItem = {
  menuItemId: number;
  quantity: number;
};

export type CreateOrderBodyPaymentMethod =
  (typeof CreateOrderBodyPaymentMethod)[keyof typeof CreateOrderBodyPaymentMethod];

export const CreateOrderBodyPaymentMethod = {
  cash: "cash",
  upi: "upi",
  card: "card",
} as const;

export interface CreateOrderBody {
  items: CreateOrderBodyItemsItem[];
  paymentMethod: CreateOrderBodyPaymentMethod;
}

export interface AnalyticsSummary {
  totalRevenue: number;
  totalOrders: number;
  averageOrderValue: number;
  period: string;
}

export interface TopItem {
  menuItemId: number;
  menuItemName: string;
  totalQuantity: number;
  totalRevenue: number;
}

export type ListOrdersParams = {
  limit?: number;
  offset?: number;
};

export type GetAnalyticsSummaryParams = {
  period?: GetAnalyticsSummaryPeriod;
};

export type GetAnalyticsSummaryPeriod =
  (typeof GetAnalyticsSummaryPeriod)[keyof typeof GetAnalyticsSummaryPeriod];

export const GetAnalyticsSummaryPeriod = {
  today: "today",
  week: "week",
  month: "month",
  all: "all",
} as const;

export type GetTopItemsParams = {
  limit?: number;
  period?: GetTopItemsPeriod;
};

export type GetTopItemsPeriod =
  (typeof GetTopItemsPeriod)[keyof typeof GetTopItemsPeriod];

export const GetTopItemsPeriod = {
  today: "today",
  week: "week",
  month: "month",
  all: "all",
} as const;