# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant Billing API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary List all menu items
 */
export const ListMenuItemsResponseItem = zod.object({
  id: zod.number(),
  name: zod.string(),
  price: zod.number(),
  category: zod.string(),
  imageUrl: zod.string(),
  available: zod.boolean(),
  createdAt: zod.string(),
});
export const ListMenuItemsResponse = zod.array(ListMenuItemsResponseItem);

/**
 * @summary Create a new menu item
 */
export const CreateMenuItemBody = zod.object({
  name: zod.string(),
  price: zod.number(),
  category: zod.string(),
  imageUrl: zod.string(),
  available: zod.boolean().optional(),
});

/**
 * @summary Get a menu item by id
 */
export const GetMenuItemParams = zod.object({
  id: zod.coerce.number(),
});

export const GetMenuItemResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  price: zod.number(),
  category: zod.string(),
  imageUrl: zod.string(),
  available: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Update a menu item
 */
export const UpdateMenuItemParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateMenuItemBody = zod.object({
  name: zod.string(),
  price: zod.number(),
  category: zod.string(),
  imageUrl: zod.string(),
  available: zod.boolean().optional(),
});

export const UpdateMenuItemResponse = zod.object({
  id: zod.number(),
  name: zod.string(),
  price: zod.number(),
  category: zod.string(),
  imageUrl: zod.string(),
  available: zod.boolean(),
  createdAt: zod.string(),
});

/**
 * @summary Delete a menu item
 */
export const DeleteMenuItemParams = zod.object({
  id: zod.coerce.number(),
});

export const DeleteMenuItemResponse = zod.object({
  success: zod.boolean(),
});

/**
 * @summary List orders
 */
export const listOrdersQueryLimitDefault = 50;
export const listOrdersQueryOffsetDefault = 0;

export const ListOrdersQueryParams = zod.object({
  limit: zod.coerce.number().default(listOrdersQueryLimitDefault),
  offset: zod.coerce.number().default(listOrdersQueryOffsetDefault),
});

export const ListOrdersResponseItem = zod.object({
  id: zod.number(),
  orderNumber: zod.string(),
  totalAmount: zod.number(),
  paymentMethod: zod.string(),
  status: zod.string(),
  items: zod.array(
    zod.object({
      menuItemId: zod.number(),
      menuItemName: zod.string(),
      quantity: zod.number(),
      unitPrice: zod.number(),
      subtotal: zod.number(),
    }),
  ),
  createdAt: zod.string(),
});
export const ListOrdersResponse = zod.array(ListOrdersResponseItem);

/**
 * @summary Create a new order (checkout)
 */
export const CreateOrderBody = zod.object({
  items: zod.array(
    zod.object({
      menuItemId: zod.number(),
      quantity: zod.number(),
    }),
  ),
  paymentMethod: zod.enum(["cash", "upi", "card"]),
});

/**
 * @summary Get order by id
 */
export const GetOrderParams = zod.object({
  id: zod.coerce.number(),
});

export const GetOrderResponse = zod.object({
  id: zod.number(),
  orderNumber: zod.string(),
  totalAmount: zod.number(),
  paymentMethod: zod.string(),
  status: zod.string(),
  items: zod.array(
    zod.object({
      menuItemId: zod.number(),
      menuItemName: zod.string(),
      quantity: zod.number(),
      unitPrice: zod.number(),
      subtotal: zod.number(),
    }),
  ),
  createdAt: zod.string(),
});

/**
 * @summary Get sales analytics summary
 */
export const getAnalyticsSummaryQueryPeriodDefault = `month`;

export const GetAnalyticsSummaryQueryParams = zod.object({
  period: zod
    .enum(["today", "week", "month", "all"])
    .default(getAnalyticsSummaryQueryPeriodDefault),
});

export const GetAnalyticsSummaryResponse = zod.object({
  totalRevenue: zod.number(),
  totalOrders: zod.number(),
  averageOrderValue: zod.number(),
  period: zod.string(),
});

/**
 * @summary Get top-selling menu items
 */
export const getTopItemsQueryLimitDefault = 5;
export const getTopItemsQueryPeriodDefault = `month`;

export const GetTopItemsQueryParams = zod.object({
  limit: zod.coerce.number().default(getTopItemsQueryLimitDefault),
  period: zod
    .enum(["today", "week", "month", "all"])
    .default(getTopItemsQueryPeriodDefault),
});

export const GetTopItemsResponseItem = zod.object({
  menuItemId: zod.number(),
  menuItemName: zod.string(),
  totalQuantity: zod.number(),
  totalRevenue: zod.number(),
});
export const GetTopItemsResponse = zod.array(GetTopItemsResponseItem);