# lib/api-zod/src/generated/types/createMenuItemBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant Billing API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMenuItemBody {
  name: string;
  price: number;
  category: string;
  imageUrl: string;
  available?: boolean;
}