# lib/api-zod/src/generated/types/createOrderBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant Billing API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateOrderBodyItemsItem } from "./createOrderBodyItemsItem";
import type { CreateOrderBodyPaymentMethod } from "./createOrderBodyPaymentMethod";

export interface CreateOrderBody {
  items: CreateOrderBodyItemsItem[];
  paymentMethod: CreateOrderBodyPaymentMethod;
}