# lib/api-zod/src/generated/types/topItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Restaurant Billing API
 * OpenAPI spec version: 0.1.0
 */

export interface TopItem {
  menuItemId: number;
  menuItemName: string;
  totalQuantity: number;
  totalRevenue: number;
}